New pride banner design
This new design removes the progress flag from the background of banner, and replace it with the normal pride flag.

This flag, including trans and bipoc colors, is controversial in the LGBT community, as it implies those people weren't included in the original flag, which doesn't make sense when [considering the origins of the movement](https://en.wikipedia.org/wiki/Stonewall_riots). For those reasons, we decided to switch the flag to the normal rainbow one.

This change comes after a few users pointing us at the controversies around this flag.

I also made the background lighter to make it clearer to see.